fix: improve test infrastructure with proper PostgreSQL setup and transaction isolation
- Refactor test configuration for better SQLModel compatibility
- Consolidate test fixtures in main conftest.py
- Use transaction isolation for tests to improve performance
- Add robust table verification in test setup
- Update run_tests.sh to support both grouped and single test runs
